+++ cprover_bindings/src/goto_program/expr.rs
@@ -198,6 +198,7 @@ pub enum BinaryOperator {
     Bitxor,
     Div,
     Equal,
+    FloatbvRoundToIntegral,
     Ge,
     Gt,
     IeeeFloatEqual,
@@ -293,10 +294,13 @@ pub fn arithmetic_overflow_result_type(operand_type: Type) -> Type {
     // give the struct the name "overflow_result_<type>", e.g.
     // "overflow_result_Unsignedbv"
     let name: InternedString = format!("overflow_result_{operand_type:?}").into();
-    Type::struct_type(name, vec![
-        DatatypeComponent::field(ARITH_OVERFLOW_RESULT_FIELD, operand_type),
-        DatatypeComponent::field(ARITH_OVERFLOW_OVERFLOWED_FIELD, Type::bool()),
-    ])
+    Type::struct_type(
+        name,
+        vec![
+            DatatypeComponent::field(ARITH_OVERFLOW_RESULT_FIELD, operand_type),
+            DatatypeComponent::field(ARITH_OVERFLOW_OVERFLOWED_FIELD, Type::bool()),
+        ],
+    )
 }
 
 ///////////////////////////////////////////////////////////////////////////////////////////////
@@ -1044,6 +1048,7 @@ impl Expr {
                     "vector comparison operators must be typechecked by `typecheck_vector_cmp_expr`"
                 )
             }
+            FloatbvRoundToIntegral => lhs.typ.is_floating_point() && rhs.typ.is_integer(),
         }
     }
 
@@ -1082,6 +1087,11 @@ impl Expr {
             VectorEqual | VectorNotequal | VectorGe | VectorLe | VectorGt | VectorLt => {
                 unreachable!(
                     "return type for vector comparison operators depends on the place type"
+                )
+            }
+            FloatbvRoundToIntegral => {
+                unreachable!(
+                    "return type for float-to-integer rounding operator depends on the place type"
                 )
             }
         }
@@ -1325,6 +1335,11 @@ impl Expr {
         assert!(!self.is_side_effect() && !e.is_side_effect());
         let cmp = self.clone().gt(e.clone());
         cmp.ternary(self, e)
+    }
+
+    /// floating-point to integer rounding
+    pub fn floatbv_round_to_integral(f: Expr, rm: Expr, ret_typ: Type) -> Expr {
+        expr!(BinOp { op: FloatbvRoundToIntegral, lhs: f, rhs: rm }, ret_typ)
     }
 }