+++ kani-driver/src/args/mod.rs
@@ -3,6 +3,7 @@
 //! Module that define Kani's command line interface. This includes all subcommands.
 
 pub mod assess_args;
+pub mod autoharness_args;
 pub mod cargo;
 pub mod common;
 pub mod list_args;
@@ -145,6 +146,8 @@ pub enum StandaloneSubcommand {
     VerifyStd(Box<std_args::VerifyStdArgs>),
     /// List contracts and harnesses.
     List(Box<list_args::StandaloneListArgs>),
+    /// Scan the input file for functions eligible for automatic (i.e., harness-free) verification and verify them.
+    Autoharness(Box<autoharness_args::StandaloneAutoharnessArgs>),
 }
 
 #[derive(Debug, clap::Parser)]
@@ -173,6 +176,9 @@ pub enum CargoKaniSubcommand {
 
     /// List contracts and harnesses.
     List(Box<list_args::CargoListArgs>),
+
+    /// Scan the crate for functions eligible for automatic (i.e., harness-free) verification and verify them.
+    Autoharness(Box<autoharness_args::CargoAutoharnessArgs>),
 }
 
 // Common arguments for invoking Kani for verification purpose. This gets put into KaniContext,
@@ -334,6 +340,10 @@ pub struct VerificationArgs {
     #[arg(long)]
     pub harness_timeout: Option<Timeout>,
 
+    /// Stop the verification process as soon as one of the harnesses fails.
+    #[arg(long)]
+    pub fail_fast: bool,
+
     /// Arguments to pass down to Cargo
     #[command(flatten)]
     pub cargo: CargoCommonArgs,
@@ -483,6 +493,7 @@ impl ValidateArgs for StandaloneArgs {
         match &self.command {
             Some(StandaloneSubcommand::VerifyStd(args)) => args.validate()?,
             Some(StandaloneSubcommand::List(args)) => args.validate()?,
+            Some(StandaloneSubcommand::Autoharness(args)) => args.validate()?,
             // TODO: Invoke PlaybackArgs::validate()
             None | Some(StandaloneSubcommand::Playback(..)) => {}
         };
@@ -528,6 +539,7 @@ impl ValidateArgs for CargoKaniSubcommand {
         match self {
             // Assess doesn't implement validation yet.
             CargoKaniSubcommand::Assess(_) => Ok(()),
+            CargoKaniSubcommand::Autoharness(autoharness) => autoharness.validate(),
             CargoKaniSubcommand::Playback(playback) => playback.validate(),
             CargoKaniSubcommand::List(list) => list.validate(),
         }