+++ kani-driver/src/main.rs
@@ -5,6 +5,7 @@ use std::ffi::OsString;
 use std::process::ExitCode;
 
 use anyhow::Result;
+use autoharness::{autoharness_cargo, autoharness_standalone};
 use time::{OffsetDateTime, format_description};
 
 use args::{CargoKaniSubcommand, check_is_valid};
@@ -22,6 +23,7 @@ use tracing::debug;
 mod args;
 mod args_toml;
 mod assess;
+mod autoharness;
 mod call_cargo;
 mod call_cbmc;
 mod call_goto_cc;
@@ -72,28 +74,33 @@ fn cargokani_main(input_args: Vec<OsString>) -> Result<()> {
         return list_cargo(*list_args, args.verify_opts);
     }
 
-    let session = session::KaniSession::new(args.verify_opts)?;
-
-    if !session.args.common_args.quiet {
-        print_kani_version(InvocationType::CargoKani(input_args));
+    if let Some(CargoKaniSubcommand::Autoharness(autoharness_args)) = args.command {
+        return autoharness_cargo(*autoharness_args);
     }
 
-    match args.command {
-        Some(CargoKaniSubcommand::Assess(args)) => {
-            return assess::run_assess(session, *args);
+    let mut session = match args.command {
+        Some(CargoKaniSubcommand::Assess(assess_args)) => {
+            let sess = session::KaniSession::new(args.verify_opts)?;
+            return assess::run_assess(sess, *assess_args);
         }
         Some(CargoKaniSubcommand::Playback(args)) => {
             return playback_cargo(*args);
         }
-        Some(CargoKaniSubcommand::List(_)) => unreachable!(),
-        None => {}
+        Some(CargoKaniSubcommand::Autoharness(_)) | Some(CargoKaniSubcommand::List(_)) => {
+            unreachable!()
+        }
+        None => session::KaniSession::new(args.verify_opts)?,
+    };
+
+    if !session.args.common_args.quiet {
+        print_kani_version(InvocationType::CargoKani(input_args));
     }
 
     if session.args.assess {
         return assess::run_assess(session, assess::AssessArgs::default());
     }
 
-    let project = project::cargo_project(&session, false)?;
+    let project = project::cargo_project(&mut session, false)?;
     if session.args.only_codegen { Ok(()) } else { verify_project(project, session) }
 }
 
@@ -103,6 +110,9 @@ fn standalone_main() -> Result<()> {
     check_is_valid(&args);
 
     let (session, project) = match args.command {
+        Some(StandaloneSubcommand::Autoharness(args)) => {
+            return autoharness_standalone(*args);
+        }
         Some(StandaloneSubcommand::Playback(args)) => return playback_standalone(*args),
         Some(StandaloneSubcommand::List(list_args)) => {
             return list_standalone(*list_args, args.verify_opts);