+++ kani-driver/src/metadata.rs
@@ -97,6 +97,7 @@ pub fn merge_kani_metadata(files: Vec<KaniMetadata>) -> KaniMetadata {
         unsupported_features: vec![],
         test_harnesses: vec![],
         contracted_functions: vec![],
+        autoharness_skipped_fns: None,
     };
     for md in files {
         // Note that we're taking ownership of the original vec, and so we can move the data into the new data structure.
@@ -106,6 +107,7 @@ pub fn merge_kani_metadata(files: Vec<KaniMetadata>) -> KaniMetadata {
         result.unsupported_features.extend(md.unsupported_features);
         result.test_harnesses.extend(md.test_harnesses);
         result.contracted_functions.extend(md.contracted_functions);
+        // We do not handle autoharness metadata here, since this function is not reachable from the autoharness subcommand.
     }
     result
 }
@@ -176,6 +178,10 @@ fn find_proof_harnesses<'a>(
     debug!(?targets, "find_proof_harness");
     let mut result = vec![];
     for md in all_harnesses.iter() {
+        // --harnesses should not select automatic harnesses
+        if md.is_automatically_generated {
+            continue;
+        }
         if exact_filter {
             // Check for exact match only
             if targets.contains(&md.pretty_name) {
@@ -224,6 +230,7 @@ pub mod tests {
             goto_file: model_file,
             contract: Default::default(),
             has_loop_contracts: false,
+            is_automatically_generated: false,
         }
     }